Split WorkspaceManagerService.java
